split model from immutable model
