+++ backend/common/infrastructure/src/database/model/mod.rs
@@ -1,11 +1,11 @@
 mod repository;
 
 use diesel::{Connection, Identifiable};
-pub use repository::Repository;
+pub use repository::{ImmutableRepository, Repository};
 
 use super::Result;
 
-pub trait Model<C>
+pub trait ImmutableModel<C>
 where
 	C: Connection,
 	Self: Identifiable + Sized,
@@ -15,8 +15,15 @@ where
 	fn list(conn: &mut C) -> Result<Vec<Self>>;
 	fn insert(self, conn: &mut C) -> Result<Self>;
 	fn try_insert(self, conn: &mut C) -> Result<Self>;
-	fn update(self, conn: &mut C) -> Result<Self>;
-	fn upsert(self, conn: &mut C) -> Result<Self>;
 	fn delete(conn: &mut C, id: <Self as Identifiable>::Id) -> Result<Self>;
 	fn clear(conn: &mut C) -> Result<()>;
+}
+
+pub trait Model<C>: ImmutableModel<C>
+where
+	C: Connection,
+	Self: Identifiable + Sized,
+{
+	fn update(self, conn: &mut C) -> Result<Self>;
+	fn upsert(self, conn: &mut C) -> Result<Self>;
 }
