+++ backend/common/infrastructure/src/database/model/repository.rs
@@ -1,26 +1,24 @@
 use diesel::{Identifiable, PgConnection};
 
-use super::{Model, Result};
+use super::{ImmutableModel, Model, Result};
 use crate::database;
 
-pub trait Repository<M>: Send + Sync
+pub trait ImmutableRepository<M>: Send + Sync
 where
-	M: Model<PgConnection>,
+	M: ImmutableModel<PgConnection>,
 {
 	fn exists(&self, id: <M as Identifiable>::Id) -> Result<bool>;
 	fn find_by_id(&self, id: <M as Identifiable>::Id) -> Result<M>;
 	fn list(&self) -> Result<Vec<M>>;
 	fn insert(&self, model: M) -> Result<M>;
 	fn try_insert(&self, model: M) -> Result<M>;
-	fn update(&self, model: M) -> Result<M>;
-	fn upsert(&self, model: M) -> Result<M>;
 	fn delete(&self, id: <M as Identifiable>::Id) -> Result<M>;
 	fn clear(&self) -> Result<()>;
 }
 
-impl<M> Repository<M> for database::Client
+impl<M> ImmutableRepository<M> for database::Client
 where
-	M: Model<PgConnection>,
+	M: ImmutableModel<PgConnection>,
 {
 	fn exists(&self, id: <M as Identifiable>::Id) -> Result<bool> {
 		let mut connection = self.connection()?;
@@ -47,16 +45,6 @@ where
 		model.try_insert(&mut *connection)
 	}
 
-	fn update(&self, model: M) -> Result<M> {
-		let mut connection = self.connection()?;
-		model.update(&mut *connection)
-	}
-
-	fn upsert(&self, model: M) -> Result<M> {
-		let mut connection = self.connection()?;
-		model.upsert(&mut *connection)
-	}
-
 	fn delete(&self, id: <M as Identifiable>::Id) -> Result<M> {
 		let mut connection = self.connection()?;
 		M::delete(&mut *connection, id)
@@ -65,5 +53,28 @@ where
 	fn clear(&self) -> Result<()> {
 		let mut connection = self.connection()?;
 		M::clear(&mut *connection)
+	}
+}
+
+pub trait Repository<M>: ImmutableRepository<M>
+where
+	M: Model<PgConnection>,
+{
+	fn update(&self, model: M) -> Result<M>;
+	fn upsert(&self, model: M) -> Result<M>;
+}
+
+impl<M> Repository<M> for database::Client
+where
+	M: Model<PgConnection>,
+{
+	fn update(&self, model: M) -> Result<M> {
+		let mut connection = self.connection()?;
+		model.update(&mut *connection)
+	}
+
+	fn upsert(&self, model: M) -> Result<M> {
+		let mut connection = self.connection()?;
+		model.upsert(&mut *connection)
 	}
 }
